+++ src/main/java/com/mycompany/CodeforSefinal/Objects/QuantityItem.java
@@ -7,13 +7,13 @@ public class QuantityItem extends Item {
     private int quantity;
     
     //Default quantity
-    public QuantityItem(long itemId, String name, double price) {
+    public QuantityItem(int itemId, String name, double price) {
         super(itemId, name, price);
         quantity = 1;
     }
     
      //Custom quantity
-    public QuantityItem(long itemId, String name, double price, int quantity) {
+    public QuantityItem(int itemId, String name, double price, int quantity) {
         super(itemId, name, price);
         this.quantity = quantity;
     }
