+++ src/main/java/com/mycompany/CodeforSefinal/Objects/ReferenceItem.java
@@ -4,7 +4,7 @@ import com.mycompany.CodeforSefinal.Objects.Item;
 
 public class ReferenceItem extends Item {
     
-    public ReferenceItem(long itemId, String name, double price) {
+    public ReferenceItem(int itemId, String name, double price) {
         super(itemId, name, price);
     }
